OOP structure (lot of singletons, probably needs review)
