+++ gulpfile.js
@@ -56,7 +56,7 @@ gulp.task('js', function () {
     var uglify = require('gulp-uglify');
 
     var bundle = browserify({
-        entries: 'src/js/main.js',
+        entries: 'src/js/tangram-play.js',
         debug: true,
         transform: [babelify]
     });
