+++ src/js/core/common.js
@@ -1,5 +1,3 @@
-'use strict';
-
 export function fetchHTTP(url, methood) {
     var request = new XMLHttpRequest(), response;
 
@@ -11,6 +9,21 @@ export function fetchHTTP(url, methood) {
     request.open(methood ? methood : 'GET', url, false);
     request.send();
     return response;
+};
+
+export function debounce(func, wait, immediate) {
+    var timeout;
+    return function() {
+        var context = this, args = arguments;
+        var later = function() {
+            timeout = null;
+            if (!immediate) func.apply(context, args);
+        };
+        var callNow = immediate && !timeout;
+        clearTimeout(timeout);
+        timeout = setTimeout(later, wait);
+        if (callNow) func.apply(context, args);
+    };
 };
 
 export function getPosition(dom) {
