+++ src/js/core/map.js
@@ -1,73 +1,71 @@
-'use strict';
-
 var take_screenshot = false;
 
-module.exports = {
-    init,
-    postUpdate
-}
+export default class Map {
+    constructor(place, style_file){
+        let map_start_location = [0.0, 0.0, 3];
 
-function init (place, style_file) {
-    var map_start_location = [0.0, 0.0, 3];
+        /*** URL parsing ***/
 
-    /*** URL parsing ***/
-
-    // leaflet-style URL hash pattern:
-    // ?style.yaml#[zoom],[lat],[lng]
-    var url_hash = window.location.hash.slice(1).split('/');
-    if (url_hash.length == 3) {
-        map_start_location = [url_hash[1],url_hash[2], url_hash[0]];
-        // convert from strings
-        map_start_location = map_start_location.map(Number);
-    }
-
-    /*** Map ***/
-    var map = L.map(place,
-        { zoomControl: false },
-        {'keyboardZoomOffset': .05}
-    );
-
-    L.control.zoom({position: 'topright'}).addTo(map);
-
-    map.attributionControl.setPrefix('<a href="http://leafletjs.com" title="A JS library for interactive maps" target="_blank">Leaflet</a>')
-    var layer = Tangram.leafletLayer({
-        scene: style_file,
-        postUpdate: postUpdate,
-        attribution: '<a href="https://mapzen.com/tangram" target="_blank">Tangram</a> | &copy; OSM contributors | <a href="https://mapzen.com/" target="_blank">Mapzen</a>'
-    });
-
-    window.layer = layer;
-    var scene = layer.scene;
-    window.scene = scene;
-
-    map.setView(map_start_location.slice(0, 2), map_start_location[2]);
-    var hash = new L.Hash(map);
-
-    window.addEventListener('load', function () {
-        // Scene initialized
-        layer.addTo(map);
-    });
-
-    //disable mousewheel zoom if iframed
-    if (window.self !== window.top) {
-        map.scrollWheelZoom.disable();
-    }
-
-    return map;
-}
-
-function postUpdate() {
-    if (take_screenshot == true) {
-        // Adapted from: https://gist.github.com/unconed/4370822
-        var image = scene.canvas.toDataURL('image/png').slice(22); // slice strips host/mimetype/etc.
-        var data = atob(image); // convert base64 to binary without UTF-8 mangling
-        var buf = new Uint8Array(data.length);
-        for (var i = 0; i < data.length; ++i) {
-            buf[i] = data.charCodeAt(i);
+        // leaflet-style URL hash pattern:
+        // ?style.yaml#[zoom],[lat],[lng]
+        let url_hash = window.location.hash.slice(1).split('/');
+        if (url_hash.length == 3) {
+            map_start_location = [url_hash[1],url_hash[2], url_hash[0]];
+            // convert from strings
+            map_start_location = map_start_location.map(Number);
         }
-        var blob = new Blob([buf], { type: 'image/png' });
-        saveAs(blob, 'tangram-' + (+new Date()) + '.png'); // uses FileSaver.js: https://github.com/eligrey/FileSaver.js/
 
-        take_screenshot = false;
-    }
-}
+        /*** Map ***/
+        let map = L.map(place,
+            { zoomControl: false },
+            {'keyboardZoomOffset': .05}
+        );
+
+        L.control.zoom({position: 'topright'}).addTo(map);
+
+        map.attributionControl.setPrefix('<a href="http://leafletjs.com" title="A JS library for interactive maps" target="_blank">Leaflet</a>')
+        let layer = Tangram.leafletLayer({
+            scene: style_file,
+            // postUpdate: this.postUpdate,
+            attribution: '<a href="https://mapzen.com/tangram" target="_blank">Tangram</a> | &copy; OSM contributors | <a href="https://mapzen.com/" target="_blank">Mapzen</a>'
+        }); 
+        map.setView(map_start_location.slice(0, 2), map_start_location[2]);
+
+        this.hash = new L.Hash(map);
+
+        window.layer = layer;
+        window.scene = layer.scene;
+
+        window.addEventListener('load', function () {
+            // Scene initialized
+            layer.addTo(map);
+        });
+
+        this.leaflet = map;
+        this.layer = layer;
+        this.scene = layer.scene;
+    };
+
+    init () {
+        this.layer.addTo(this.leaflet);
+    };
+
+    postUpdate () {
+        if (take_screenshot == true) {
+            // Adapted from: https://gist.github.com/unconed/4370822
+            var image = scene.canvas.toDataURL('image/png').slice(22); // slice strips host/mimetype/etc.
+            var data = atob(image); // convert base64 to binary without UTF-8 mangling
+            var buf = new Uint8Array(data.length);
+            for (var i = 0; i < data.length; ++i) {
+                buf[i] = data.charCodeAt(i);
+            }
+            var blob = new Blob([buf], { type: 'image/png' });
+            saveAs(blob, 'tangram-' + (+new Date()) + '.png'); // uses FileSaver.js: https://github.com/eligrey/FileSaver.js/
+
+            take_screenshot = false;
+        }
+    };
+};
+
+
+
